requirements.txt
accelerate==0.15.0
boto3==1.26.93
botocore==1.29.94
chainer==7.8.1
clip==0.2.0
cupy-cuda111==8.1.0
deepspeed==0.7.7
diffusers==0.10.2
einops==0.6.0
ftfy==6.1.1
gdown==4.6.0
gifmaker==1.5
huggingface-hub==0.11.1
imageio==2.25.0
matplotlib==3.7.0
mmcv-full==1.7.0
mmdet==2.25.3
multiprocess==0.70.14
numpy==1.23.5
omegaconf==2.3.0
open-clip-torch==2.16.0
opencv-python==4.7.0.68
pandas==1.5.3
Pillow==9.3.0
pyparsing==3.0.9
rotary-embedding-torch==0.1.5
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
taming-transformers==0.0.1
tensorboard==2.11.0
timm==0.6.12
tokenizers==0.13.2
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchmetrics==0.11.1
torchvision==0.13.0+cu113
tqdm==4.64.1
transformers==4.25.1
xformers==0.0.13
